moved components into appropriate dir
